RavenDB-12581 Windows journal functions moved to one file
